DONE: default sorting feature implemented; TODO: transplant features between Views
