+++ CardBox/CardBox/SubViews/AddButtonView.swift
@@ -1,0 +1,22 @@
+//
+//  AddButtonView.swift
+//  CardBox
+//
+//  Created by Minjong Ha on 2022/03/26.
+//
+
+import SwiftUI
+
+struct AddButtonView: View {
+    var body: some View {
+        NavigationLink(destination: OnDemandView(AddNewCardView())) {
+            Text("Add")
+        }
+    }
+}
+
+struct AddButtonView_Previews: PreviewProvider {
+    static var previews: some View {
+        AddButtonView()
+    }
+}
