+++ CardBox/CardBox/Utils/Card.swift
@@ -8,7 +8,6 @@
 import Foundation
 import RealmSwift
 
-//TODO: add func() for Object interaction?
 //TODO: update schema caution (https://docs.mongodb.com/realm/sdk/swift/examples/modify-an-object-schema/)
 struct CardCell: Hashable {
     var cardUUID: String = "" // primary key
